Add User Integration Spec
