+++ app/views/user/show.html.erb
@@ -1,4 +1,4 @@
-  
+ <h4>User Details</h4> 
 <div class="card mb-3">
   <div class="row g-0">
     <div class="col-md-4">
