+++ config/routes.rb
@@ -1,6 +1,6 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-  get '/', to: 'users#index'
+  get '/', to: 'user#index'
 
   resources :user, only: %i[index show] do
     resources :post, only: %i[index show new create] do
